AUT-4633: Emit new metric from NotificationHandler
Call the CloudWatchService to increment the new SmsLimitExceeded
metric when a 429 response is received from Notify.

We are using 429 responses as the signal that the quota has been
exceeded because the "the limit within Notify is being adjusted fairly
frequently at present in response to spiking demand from spam
accounts".  This approach will be kept under review.