BAU: Switch off Dynamo streaming on user-profile in prod environments
Dynamo streaming was originally used to trigger an Experian phone check when a phone number was added to the database.  This is now deprecated and has been replaced by an SQS message instead, but streaming is still switched on.

Following the switch-off in test environments this change switches off in integration and production.