AUT-4557: Support new fields in email check result messages
The check service returns extra data about the check result now, and we need to record that data to raise in a later `*_DECISION_USED` audit event.

The content of `emailCheckResponse` is explicitly not defined as to hide the structure of it's content. The content is already provided in the structure necessary for the audit event.