AUT-4557: Allow new data recording in cache
Update the EmailCheckResultStore to include the new journey ID and EmailCheckResponse field data. This will be used in the `_DECISION_USED` audit event later.

The `emailCheckResponse` field is explicitly ambiguous as not to expose the structure of its content. It is already provided in the structure needed for the audit event.

As such, a new ObjectConverter class is included to translate to and from JSON and Java Objects.