AUT-4572: Consolidate LoginHandler Password lockout integration tests
We had 2 existing tests that do almost the exact same thing, just assert differently. We also have a gap that does not cover checking the lockout is still in place when trying to enter password again after being lockout out.

So here we consolidate and add to cover these gaps.

This is in preparation to refactor the lockout business logic.