AUT-4572: Tweak AuditEventExpectation
Rather than passing in the eventName string, now construct the expectation from the AuditableEvent itself. This is useful for an upcoming test helper that will just accept a full set of AuditEventExpectations and assert they are in the queue without having to do additional conversions later.

verify is also updated to not pick the first receivedEvent that matches the expected event name. This does not work when there are multiple expectations for the same event name but different properties. Instead we check all events (exiting early for name mismatches) to assert the received event matches the expectation.