AUT-4572: Tweak PermissionDecisionManager constructor
In preparation for adding the AuthenticationAttemptsService to the PermissionDecisionManager, tweak the existing constructors to simplify the interface. At a minimum, the ConfigurationService should be passed in (no longer using the singleton method to grab the dependency.

The handler constructors updated here are only used by tests, not when executed in a lambda.