ATO-1875: Add method to get list from secrets manager
Adds a method to fetch the allow list from secrets manager. This
does some sensible 'failing-open' in the case that we encounter
a ResourceNotFound exception by returning an empty list - effectively
meaning that no test journey is permitted.

We do the same in the case the secretString is null or empty.

I intended to keep this class and methods static - however the
ease of mocking secrets manager by depedency injection made me
abadon this.

There is room to improve this fetching by caching the list - however
I have not implemented this here for simplicty and due to the fact
that the rate limit of GetSecretValue[1] is quite high. If this is
seen to be a bottleneck in future this can easily be implemented