AUT-4572: Ensure canReceivePassword returns the correct attemptCount
The intent of the UserPermissionsManager is to centralise and align the lockout business logic accessing different data storage locations.

Non-reauth journeys store some counts in the codeStorageService, that get reset when a block record is written.

All Decisions return an attempt count, but for canReceivePassword on a non-reauth journey this will be 0 when a block is in place.

Instead of that, align the counts so that the real count that caused the block is returned.

This is used later to tidy up the logic in LoginHandler in regards to keeping track of the attempt count.