AUT-4572: Create getReauthAttemptCounts in LoginHandler
We have a situation where we may be unnecessarily adding all reauth lockout counts to an audit event where they are optional. While checks are taking place to assert if we can simplify this the getReauthAttemptCounts is created.

It pulls the count fetch into a shared method in LoginHandler rather than expand the scope of the UserPermissionsManager for the time being. If we can simplify the logic then we can populate the audit event with the data already returned by the UserPermissionsManager.

If we need to keep existing functionality we will need to expand UserPermissionsManager in order to remove the direct data storage access from the handler.