AUT-4572: Remove canVerifySmsOtp and canVerifyAuthAppOtp
This consolidates the intended functionality into `canVerifyMfaOtp`

The original intent of these separate methods was to make an explicit distinction between the user initiated action and the lockout behind it. In other words, we treat the lockout and count for verifying SMS and AUTH_APP OTPs as the same, but intended for the `PermissionDecisionManager` not to leak this lockout implementation detail via it's API by using a combined `canVerifyMfaOtp` method.

However, using two distinct methods is not performant or introduces code smells. The same check will either be done twice, or we would need conditional logic in the handlers to decide which method to call.

Instead, let's just use a simplified API of `canVerifyMfaOtp`