ATO-1884: Implement simple secret cache
AWS recommend that you cache the response of a GetSecretValue command[1]
to reduce the cost of calls to secrets manager. We should implement this
to ensure we don't end up exceeding the rate limit of calls to secrets manager
under high load.

I have tried to implement this in a way that does not conflict with using snapstart
so the cache has an expiry of 300 seconds and is not initalised in the constructor
so will not be part of snapstarts initlization phase[2]

[1]- https://docs.aws.amazon.com/secretsmanager/latest/userguide/retrieving-secrets-java.html#:~:text=However%2C%20we%20recommend%20that%20you%20cache%20your%20secret%20values%20by%20using%20client%2Dside%20caching.%20Caching%20secrets%20improves%20speed%20and%20reduces%20your%20costs.
[2] - https://docs.aws.amazon.com/lambda/latest/dg/snapstart.html#:~:text=Temporary%20data,not%20using%20SnapStart.