ATO-1884: Add addtional tests for test client cases
As part of previous changes[1], the sendOtpNotification handler
started writing notify messages on the SQS queue for these to be
written to S3. This requires re-jigging the tests to now assert
that values are still written to the queue but with a flag to
indicate these are test user requests.

[1] -https://github.com/govuk-one-login/authentication-api/pull/6281