AUT-4572: Instantiate CodeStorageService when needed
Previously a `dynamodb:DescribeTable` permission was needed to the `authentication-attempts` table for every handler instantiating the `UserActionsManager`, even if it didn't use it. So we changed it to only instantiate the `AuthenticationAttemptsService` when needed and cache the result.

We follow a similar approach here for the `CodeStorageService` to establish this pattern: "Only instantiate a storage provider if you need to access it".