AUT-4574: Switch initial lockout count identifier
This handler is responsible for checking the email address the user submits matches the email address associated with the user profile the RP requested do a reauthentication.

If it does not match it will record a lockout count increment against the rpPairwiseId the RP requested. If the user still has an active session from a previous authentication then the subject ID of the user profile of the signed in session is used as the identifier instead.

So a signed in user can keep entering the incorrect email address until they are locked out. In this scenario the lockout is associated with the signed in user profile.

Before we even get to checking if the email address submitted matches the RP expectation we check for lockouts. This happens with the `throwLockedExceptionAndEmitAuditEventIfExistentUserIsLocked` method.

This method looks up the locks by the email address sent to it and the rpPairwiseId from the RP.

However previously this was using the email address the user submits, rather than the email address of the signed in user profile. This is important as we are not recording lockouts against the email address the user submits.

In essence this is ineffective. The only scenario this would encounter is if the user profile of the email address submitted by the user happens to have a lockout at the same time this handler processes the same user submitted email (i.e. it would take 2 users doing reauthentication where user A locks themselves out while having an active session and user B submits the email address of user A).

So this commit corrects this. Instead we are checking for lockouts for the email address associated with the signed in user profile.