ATO-1866: Remove usages of DefaultValue in FindInMap
This is a pre-requisite to removing the usages of `AWS::LanguageExtensions`
as this is one of the only features of this transformation we use. This was
introduced in this [commit](5cc1f68f254a77556d99a31cad005a01328bacea)
as a way of setting the concurrency of the trustmark function lower
than the default value. This is because the trustmark function is
not hit in any user facing journeys and is just used to advertise the
supported vectors of trust, as specified in the RFC[1].

Now we have removed provisioned concurrency from the TrustMark
lambda we can remove the complexity here and remove a usage
of this extension

[1] - https://datatracker.ietf.org/doc/html/rfc8485#section-5