ATO-1866: Remove AWS::LanguageExtensions
We are seeing an issue where during promotion of our build artefact
the keys are sorted in the YAML. Normally YAML is order agnostic
but the AWS::LanguageExtensions has an issue where it expects a
specific ordering of the keys in the IAM statement.

This causes a deployment error when attempting to deploy the
template which includes the resource policy to deny access to
non approvedAdmin roles.

As an attempt to fix this, I am removing the usage 'AWS::LanguageExtensions'
transform macro. This was introduced in [commit](5cc1f68f254a77556d99a31cad005a01328bacea)
but we only really used it to apply different provisioned concurrency config
to the trustmark function. I have removed this and cannot see any other
usages of macros that require this extension so we should be good to
remove this.