ATO-1866: Remove Conditional Update/Deletion Policies
These are not supported in native cloudformation without the
use of the AWS::LanguageExtension transform.[1] Given we are removing
this transfrom - I am setting these to be retain regardless of env.

We should potentially come back to this in the future and see if
we can apply this conditionally to exclude dev. On balance though
it should be fine

[1]- https://aws.amazon.com/blogs/mt/introducing-new-language-extensions-in-aws-cloudformation/#:~:text=Intrinsic%20support%20for%20DeletionPolicy%20and%20UpdateReplacePolicy