Revert "ATO-1703: Hardcode ipv stub domain"
This reverts commit 177988352c11cb46eb2813ec269ead36b5906947.

We previously introduced this to enable the fact that we were
using authdev3 on the main API but the stubs remained on the
sandpit domain. These have been fully migrated now so we can
revert this commit.

This has the added benefit of resolving
a deployment issue with the previous changes as the value
ipvStubDomain was only present in the dev and build mapping
blocks (as it is unused everywhere else!). This must have
been supported when using the LanguageExtensions transform
but seems to now fail.