BAU: Correct TICF call headers
The call to TICF should have included a content type header but the
production TICF API doesn't enforce this.

When updating the PACT consumer tests to include payloads the PACT
framework automatically adds a content type header which makes the
tests fail.  To allow the PACTs to be defined correctly the TICF
handler was updated to send the appropriate content type header.