AUT-4834: Update AM openapi spec for SFAD
This spec version is being bumped to 3.1.0 as changes have
been made for work on the Single Factor Account Deletion
initiative. This includes:

- New send-otp-challenge endpoint, to be called by Account
  Management Components (AMC)
    - This was required as SFAD requires the OTP to be sent
      to an MFA method already associated with an account,
      which was not easily implementable with the existing
      send-otp-notification endpoint
- New verify-otp-challenge endpoint
    - This validates an OTP without performing account
      updates. Existing OTP validation methods were
      integrated into handlers (e.g. UpdateEmailHandler)
- Updated authenticate endpoint spec
    - Now reflects the pre-existing state of the API, with
      the addition of accepting a clientId query parameter
- Updated delete-account endpoint spec
    - Now reflects the pre-existing state of the API, with
      the addition of accepting a clientId query parameter

All of the aforementioned endpoints accept the clientId as
a query parameter. This is so that the
AuthoriseAccessTokenHandler can know where to fetch the
JWKS from to verify the JWTs used as bearer tokens. For
example, in the current instance, we will fetch the JWKS
from the Orchestration .well-known endpoint for the
existing account management flows, as they provision the
access token for Home. For the new SFAD flow, if the
clientId belongs to accountManagementComponents, we will
then know that the access token being returned to use is
the token we provisioned, and thus can be verified using
our public signing key.

This commit also adds changes so that the API endpoints
can be hit with desired responses when using imposter.
See rest-plugin-config to see specific endpoints to hit
to get the different responses.