ATO-1913: Add feature flag to template
Adds the feature flag to the template file, false in all envs.
Given how widely used the Client Registry is in the Orch codebase
I have made this a global env var to ensure it is applied to all
functions. This is fine as it will be ignored where not needed