ATO-2082: Remove DynamoClientService from Processing Identity Handler
As part of moving the Client Registry into the Orchestration account
we would like to not have a cross account read. Whilst the current
table resource policy supports this - it adds complexity in how we
manage code changes with the rollout of where reads/writes occur
given these will be deployed in two different pipelines.

To mitigate this - we can remove the reading of the Client Registry
here based on a couple of things:

1 - The reads were only using the client id from the client - which
we already have from the client sessio (in fact is used to fetch it
in BaseFrontendHandler)

2- We are extremely unlikely to run into the one behavioural change
in this code (throwing for no client found) - looking at the logs
I can't really ever see this occuring (in fact for this to happen
we'd need to remove a client from the registry in the within a session
expiry - very unlikley!)