BAU: Correct Token Integration Tests
Context: the createAuthCodeForClient method is returning a value that is not being used. It is called by two of the tests to save an authcode with a different clientId, to test that 400 is returned if the client id in the request doesn't match the client id saved with the auth code. The tests were passing with a 400 response by coincidence, because the default test setup was saving auth code with a different client id from the one used in the test.
- separate the steps in the constructBaseTokenRequest method to make it easier and clearer when a test needs auth code to be saved with a different client id.
- adjust the two tests in question so that the auth code is explicitly set up with a different client id from the one used in the request.
- assert on logging, to make the cause of the 400 more explicit in the test.