BAU: Fix reference to orch env in table policy
Authdev3 is a sub environment and is the only sub env that orch is
integrated with. However the corresponding orch env is dev not
a sub env. The way this was constucting the polciy in authdev3
was to use the subenv prefix for the orch tables which was incorrect
and therefore did not work