ATO-2080: Bootstrap ClientRegistry Migration handler
Adds the skeleton IaC and code for a handler which will be manually
invoked in the console (once hopefully!) to transfer data between
the Auth account CR and into the one in the Orch account.

Currently setup with all the logging/alerting we have for a usual
lambda - I left out the deployment preference from this as it is
almost certainly un-needed.

I also stripped back most of the policies to ensure this can only
read/write to the CR tables.