ATO-2080: Bootstrap ClientRegistry Migration handler
Adds the skeleton IaC and code for a handler which will be manually
invoked in the console (once hopefully!) to transfer data between
the Auth account CR and into the one in the Orch account.

I stripped back most of the policies to ensure this can only
read/write to the CR tables. I also have removed most of the
alerting we have on our lambdas as we're likley to be monitoring
this in the console when it is running