ATO-2080: Add client registry migration service
This service is a little different from our normal way of writing
to DynamoDB as it uses the 'raw' Dynamo attribute maps instead of
mapping using the enhanced client.

This is because I found that during testing, using the enhanced client
modified the data that was being read out and written to the client
registry by populating fields with default values if they were missing.

This is the intended behaviour at runtime of how we use the client registry
but perhaps not something we want to do when migrating the data between tables,
where we would like to be sure of the integrity of the data between the tables.

This service has two methods one to scan all items from a table and
another to put the item to the table - these will be used seperately
to read all the data from one table and write to the other. This is why
the constuctor has a flag to enable this