ATO-1915: Sort clients and their keys before hashing
Hashing is order dependant - and we found that during testing
we logged different hashes when comparing what looked to be
identical contents of each client registry.

This fixes the issue by sorting the client and it's keys to ensure
we hash in determined order - ensuring our hash output is effectively
only comparing the contents not the ordering it was delivered in!