AUT-4574: Fix IDs used to check for lockouts on email submission
This handler is responsible for checking the email address the user submits matches the email address associated with the user profile the RP requested do a reauthentication.

If it does not match it will record a lockout count increment against the `rpPairwiseId` the RP requested. If the user still has an active session from a previous authentication then the subject ID of the user profile of the signed in session is used as the lockout identifier instead.

So a signed in user can keep entering the incorrect email address until they are locked out. In this scenario the lockout is associated with the signed in user profile.

Previously, before we even get to checking if the email address submitted matches the RP expectation we check for lockouts. This happens with the `throwLockedExceptionAndEmitAuditEventIfExistentUserIsLocked` method.

This method looks up the locks by the email address sent to it and the rpPairwiseId from the RP.

However previously this was using the email address the user submits, rather than the email address of the signed in user profile. This is important as we are not recording lockouts against the email address the user submits.

In essence this is ineffective. The only scenario this would encounter is if the user profile of the email address submitted by the user happens to have a lockout at the same time this handler processes the same user submitted email (i.e. it would take 2 users doing reauthentication where user A locks themselves out while having an active session and user B submits the email address of user A).

So this commit corrects this.

Before we can check for the lockouts, we need to check if there is a match between the generated pairwiseId for the user profile with the same email address that the user submitted and the pairwiseId the RP submitted. So the lockout check is moved into the `verifyReAuthentication` method.

Once we know if there is a match or not we can check for lockouts against up to 3 identifiers:
- We always look for lockouts against the rpPairwiseId.
- If the user is signed in we also look for lockouts against that user.
- If there is a match then we also look for lockouts against the matched user.

The user profiles we check may not necessarily be the same. In theory the user could be signed in to auth with a different user profile than the one the RP is wanting to reauthenticate. We should be checking for lockouts against both users in that case.

Extensive testing is added in the form of the `emailSubmittedLockoutCheckScenarios` test factory which tests 24 different combinations and their expectations.