AUT-4574: Move logic from helper methods
The helper methods verifyReAuthentication and throwLockedExceptionAndEmitAuditEventIfExistentUserIsLocked do very specific lockout checks and also some duplication of count checks. Some of this duplication may be able to consolidate.

So going forward we'll move the lockout business logic out of the handler and consider creating new helper methods that may be more appropriate for the new approach