AUT-4574: Allow multiple internalSubjectIds
With the CheckReAuthUserHandler needing to support looking up lockouts for multiple internalSubjectIds (both a possible signed in user and the expected reauth user, which may not be the same), UserPermissionContext is extended to support referring to two internalSubjectIds.

Existing gets are still supported by the `internalSubjectId` getter which will throw if there are too many ids.

In addition, existing instantiations of UserPermissionContext are replaced with the builder pattern to:
- Make it easier to understand
- Remove the need to hint like this `new UserPermissionContext((String) null, null, null, null)` when the subject is null (without the hint the correct constructor cannot be deduced by the compiler)