BAU: Fix domestic SmsLimitExceeded metric
Note that this was previously using "GB" to determine if the SMS destination type was domestic or international. The maybeGetCountry helper returns the numeric country code, rather than the alphabetic.

This previously meant that the destination type would always be "INTERNATIONAL" if this alarm was triggered, which is incorrect for domestic cases.

The tests have also been fixed (there was previously a domestic and international mismatch), and an additional test put in place to ensure both the "DOMESTIC" and "INTERNATIONAL" cases have full coverage.