BAU: Add DynamoExtension to check for invalid fields
As our Client Registry class has all getters annotated with
the DynamoDbAttribute annotation, we use this to create a list
of allowed client registry field names - as any fields not
represented by this annotation will be ignored at runtime

To do this we need to use a DynamoDbEnhancedClientExtension as
we need to hook into the "raw" dynamo map object before it is mapped
to our data class. Otherwise the unknown fields are dropped. We can
then just iterate over the object keys and log any that aren't part
of our valid list fields